[tracing] Use OT provided mechanisms for configuring OT
Notably, we now use OpenTelemetry's own reflection-based mechanism for
loading tracers. This means we have removed our custom code for
setting things up for Jaeger.
