+++ java/server/src/org/openqa/selenium/grid/ActualMain.java
@@ -1,0 +1,2 @@
+package org.openqa.selenium.grid;public class ActualMain {
+}
